requirements.txt
aioredis==1.3.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.11.28
cffi==1.13.2
channels==2.3.1
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
crispy-forms-foundation==0.7.1
cryptography==2.8
daphne==2.4.0
Django==2.2.5
django-crispy-forms==1.8.1
hiredis==1.0.1
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.2
Pillow==6.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
pytz==2019.2
requests==2.22.0
service-identity==18.1.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
Twisted==19.10.0
txaio==18.8.1
urllib3==1.25.7
webencodings==0.5.1
zope.interface==4.7.1
